feat : ssr 없는 버전 제작
