+++ src/components/domains/mypage/MyPageProfile/index.tsx
@@ -25,21 +25,18 @@ interface ProfileProps {
 
 export default function MyPageProfile({ memberData }: ProfileProps) {
   const [profileEdit, setProfileEdit] = useState<boolean>(false);
-  const [nick, setNick] = useState<string>(memberData?.nickname || "");
-  const [img, setImg] = useState<string>(memberData?.imageUrl || "");
 
   const handleProfileEdit = () => {
     setProfileEdit((prev) => !prev);
   };
-
-  const profileImg = img !== null ? img : defaultImg;
+  const profileImg = memberData?.imageUrl ? memberData?.imageUrl : defaultImg;
 
   return (
     <>
       <div className={cn("profileWrap")}>
         <div className={cn("imgEditWrap")} onClick={handleProfileEdit}>
           <div className={cn("profileImg")}>
-            <Image src={profileImg} alt="프로필 이미지" fill />
+            <Image src={memberData?.imageUrl} alt="프로필 이미지" fill />
           </div>
           <div className={cn("profileEdit")}>
             <div className={cn("profileEditimg")}>
@@ -49,7 +46,7 @@ export default function MyPageProfile({ memberData }: ProfileProps) {
         </div>
 
         <div className={cn("profileName")}>
-          {nick.slice(0, 16)}
+          {memberData?.nickname.slice(0, 16)}
           {memberData?.role === "ROLE_ADMIN" && (
             <Image
               src={adminMark}
